refactor on POC
